+++ src/abcd_graph/callbacks/abstract.py
@@ -25,7 +25,7 @@ from typing import Optional
 
 from abcd_graph.exporter import GraphExporter
 from abcd_graph.graph.core.abcd_objects import GraphImpl
-from abcd_graph.graph.models import Model
+from abcd_graph.models import Model
 from abcd_graph.params import ABCDParams
